zshrc fixes for Linux
Rename zshrc to .zshrc so zsh will actually load it.
The Linux console doesn't understand the escape codes "\e[4 q" and "\e[6 q".
These codes change the cursor from and underline to a beam depending on whether
we're in insert mode or command mode. Add a check for $TERM == "linux" so that
we use different escapes when we're in the console.